+++ WynnNetSDK/Models/Leaderboard/Metadata/LeaderboardMetadataTotalLevelXpPlaytime.cs
@@ -11,17 +11,17 @@ public class LeaderboardMetadataTotalLevelXpPlaytime : LeaderboardMetadata
     /// The total combined level across all characters.
     /// </summary>
     [JsonPropertyName("totalLevel")]
-    public int TotalLevel { get; set; }
+    public ulong TotalLevel { get; set; }
 
     /// <summary>
     /// The experience points earned.
     /// </summary>
     [JsonPropertyName("xp")]
-    public int Xp { get; set; }
+    public ulong Xp { get; set; }
 
     /// <summary>
     /// The total playtime in minutes.
     /// </summary>
     [JsonPropertyName("playtime")]
-    public int Playtime { get; set; }
+    public double Playtime { get; set; }
 }
