+++ WynnNetSDK/Models/Leaderboard/Metadata/LeaderboardMetadataXpPlaytime.cs
@@ -11,11 +11,11 @@ public class LeaderboardMetadataXpPlaytime : LeaderboardMetadata
     /// The experience points earned.
     /// </summary>
     [JsonPropertyName("xp")]
-    public int Xp { get; set; }
+    public ulong Xp { get; set; }
 
     /// <summary>
     /// The total playtime in minutes.
     /// </summary>
     [JsonPropertyName("playtime")]
-    public int Playtime { get; set; }
+    public double Playtime { get; set; }
 }
